+++ printer.js
@@ -1,734 +1,10 @@
-module.exports =
-/******/ (function(modules) { // webpackBootstrap
-/******/ 	// The module cache
-/******/ 	var installedModules = {};
-
-/******/ 	// The require function
-/******/ 	function __webpack_require__(moduleId) {
-
-/******/ 		// Check if module is in cache
-/******/ 		if(installedModules[moduleId])
-/******/ 			return installedModules[moduleId].exports;
-
-/******/ 		// Create a new module (and put it into the cache)
-/******/ 		var module = installedModules[moduleId] = {
-/******/ 			exports: {},
-/******/ 			id: moduleId,
-/******/ 			loaded: false
-/******/ 		};
-
-/******/ 		// Execute the module function
-/******/ 		modules[moduleId].call(module.exports, module, module.exports, __webpack_require__);
-
-/******/ 		// Flag the module as loaded
-/******/ 		module.loaded = true;
-
-/******/ 		// Return the exports of the module
-/******/ 		return module.exports;
-/******/ 	}
-
-
-/******/ 	// expose the modules object (__webpack_modules__)
-/******/ 	__webpack_require__.m = modules;
-
-/******/ 	// expose the module cache
-/******/ 	__webpack_require__.c = installedModules;
-
-/******/ 	// __webpack_public_path__
-/******/ 	__webpack_require__.p = "";
-
-/******/ 	// Load entry module and return exports
-/******/ 	return __webpack_require__(0);
-/******/ })
-/************************************************************************/
-/******/ ([
-/* 0 */
-/***/ function(module, exports, __webpack_require__) {
-
-	module.exports = __webpack_require__(1);
-
-
-/***/ },
-/* 1 */
-/***/ function(module, exports, __webpack_require__) {
-
-	'use strict';
-
-	Object.defineProperty(exports, "__esModule", {
-	  value: true
-	});
-	exports.print = print;
-
-	var _visitor = __webpack_require__(2);
-
-	/**
-	 * Converts an AST into a string, using one set of reasonable
-	 * formatting rules.
-	 */
-	function print(ast) {
-	  return (0, _visitor.visit)(ast, { leave: printDocASTReducer });
-	} /**
-	   *  Copyright (c) 2015, Facebook, Inc.
-	   *  All rights reserved.
-	   *
-	   *  This source code is licensed under the BSD-style license found in the
-	   *  LICENSE file in the root directory of this source tree. An additional grant
-	   *  of patent rights can be found in the PATENTS file in the same directory.
-	   */
-
-	var printDocASTReducer = {
-	  Name: function Name(node) {
-	    return node.value;
-	  },
-	  Variable: function Variable(node) {
-	    return '$' + node.name;
-	  },
-
-	  // Document
-
-	  Document: function Document(node) {
-	    return join(node.definitions, '\n\n') + '\n';
-	  },
-
-	  OperationDefinition: function OperationDefinition(node) {
-	    var op = node.operation;
-	    var name = node.name;
-	    var varDefs = wrap('(', join(node.variableDefinitions, ', '), ')');
-	    var directives = join(node.directives, ' ');
-	    var selectionSet = node.selectionSet;
-	    // Anonymous queries with no directives or variable definitions can use
-	    // the query short form.
-	    return !name && !directives && !varDefs && op === 'query' ? selectionSet : join([op, join([name, varDefs]), directives, selectionSet], ' ');
-	  },
-
-
-	  VariableDefinition: function VariableDefinition(_ref) {
-	    var variable = _ref.variable,
-	        type = _ref.type,
-	        defaultValue = _ref.defaultValue;
-	    return variable + ': ' + type + wrap(' = ', defaultValue);
-	  },
-
-	  SelectionSet: function SelectionSet(_ref2) {
-	    var selections = _ref2.selections;
-	    return block(selections);
-	  },
-
-	  Field: function Field(_ref3) {
-	    var alias = _ref3.alias,
-	        name = _ref3.name,
-	        args = _ref3.arguments,
-	        directives = _ref3.directives,
-	        selectionSet = _ref3.selectionSet;
-	    return join([wrap('', alias, ': ') + name + wrap('(', join(args, ', '), ')'), join(directives, ' '), selectionSet], ' ');
-	  },
-
-	  Argument: function Argument(_ref4) {
-	    var name = _ref4.name,
-	        value = _ref4.value;
-	    return name + ': ' + value;
-	  },
-
-	  // Fragments
-
-	  FragmentSpread: function FragmentSpread(_ref5) {
-	    var name = _ref5.name,
-	        directives = _ref5.directives;
-	    return '...' + name + wrap(' ', join(directives, ' '));
-	  },
-
-	  InlineFragment: function InlineFragment(_ref6) {
-	    var typeCondition = _ref6.typeCondition,
-	        directives = _ref6.directives,
-	        selectionSet = _ref6.selectionSet;
-	    return join(['...', wrap('on ', typeCondition), join(directives, ' '), selectionSet], ' ');
-	  },
-
-	  FragmentDefinition: function FragmentDefinition(_ref7) {
-	    var name = _ref7.name,
-	        typeCondition = _ref7.typeCondition,
-	        directives = _ref7.directives,
-	        selectionSet = _ref7.selectionSet;
-	    return 'fragment ' + name + ' on ' + typeCondition + ' ' + wrap('', join(directives, ' '), ' ') + selectionSet;
-	  },
-
-	  // Value
-
-	  IntValue: function IntValue(_ref8) {
-	    var value = _ref8.value;
-	    return value;
-	  },
-	  FloatValue: function FloatValue(_ref9) {
-	    var value = _ref9.value;
-	    return value;
-	  },
-	  StringValue: function StringValue(_ref10) {
-	    var value = _ref10.value;
-	    return JSON.stringify(value);
-	  },
-	  BooleanValue: function BooleanValue(_ref11) {
-	    var value = _ref11.value;
-	    return JSON.stringify(value);
-	  },
-	  NullValue: function NullValue() {
-	    return 'null';
-	  },
-	  EnumValue: function EnumValue(_ref12) {
-	    var value = _ref12.value;
-	    return value;
-	  },
-	  ListValue: function ListValue(_ref13) {
-	    var values = _ref13.values;
-	    return '[' + join(values, ', ') + ']';
-	  },
-	  ObjectValue: function ObjectValue(_ref14) {
-	    var fields = _ref14.fields;
-	    return '{' + join(fields, ', ') + '}';
-	  },
-	  ObjectField: function ObjectField(_ref15) {
-	    var name = _ref15.name,
-	        value = _ref15.value;
-	    return name + ': ' + value;
-	  },
-
-	  // Directive
-
-	  Directive: function Directive(_ref16) {
-	    var name = _ref16.name,
-	        args = _ref16.arguments;
-	    return '@' + name + wrap('(', join(args, ', '), ')');
-	  },
-
-	  // Type
-
-	  NamedType: function NamedType(_ref17) {
-	    var name = _ref17.name;
-	    return name;
-	  },
-	  ListType: function ListType(_ref18) {
-	    var type = _ref18.type;
-	    return '[' + type + ']';
-	  },
-	  NonNullType: function NonNullType(_ref19) {
-	    var type = _ref19.type;
-	    return type + '!';
-	  },
-
-	  // Type System Definitions
-
-	  SchemaDefinition: function SchemaDefinition(_ref20) {
-	    var directives = _ref20.directives,
-	        operationTypes = _ref20.operationTypes;
-	    return join(['schema', join(directives, ' '), block(operationTypes)], ' ');
-	  },
-
-	  OperationTypeDefinition: function OperationTypeDefinition(_ref21) {
-	    var operation = _ref21.operation,
-	        type = _ref21.type;
-	    return operation + ': ' + type;
-	  },
-
-	  ScalarTypeDefinition: function ScalarTypeDefinition(_ref22) {
-	    var name = _ref22.name,
-	        directives = _ref22.directives;
-	    return join(['scalar', name, join(directives, ' ')], ' ');
-	  },
-
-	  ObjectTypeDefinition: function ObjectTypeDefinition(_ref23) {
-	    var name = _ref23.name,
-	        interfaces = _ref23.interfaces,
-	        directives = _ref23.directives,
-	        fields = _ref23.fields;
-	    return join(['type', name, wrap('implements ', join(interfaces, ', ')), join(directives, ' '), block(fields)], ' ');
-	  },
-
-	  FieldDefinition: function FieldDefinition(_ref24) {
-	    var name = _ref24.name,
-	        args = _ref24.arguments,
-	        type = _ref24.type,
-	        directives = _ref24.directives;
-	    return name + wrap('(', join(args, ', '), ')') + ': ' + type + wrap(' ', join(directives, ' '));
-	  },
-
-	  InputValueDefinition: function InputValueDefinition(_ref25) {
-	    var name = _ref25.name,
-	        type = _ref25.type,
-	        defaultValue = _ref25.defaultValue,
-	        directives = _ref25.directives;
-	    return join([name + ': ' + type, wrap('= ', defaultValue), join(directives, ' ')], ' ');
-	  },
-
-	  InterfaceTypeDefinition: function InterfaceTypeDefinition(_ref26) {
-	    var name = _ref26.name,
-	        directives = _ref26.directives,
-	        fields = _ref26.fields;
-	    return join(['interface', name, join(directives, ' '), block(fields)], ' ');
-	  },
-
-	  UnionTypeDefinition: function UnionTypeDefinition(_ref27) {
-	    var name = _ref27.name,
-	        directives = _ref27.directives,
-	        types = _ref27.types;
-	    return join(['union', name, join(directives, ' '), '= ' + join(types, ' | ')], ' ');
-	  },
-
-	  EnumTypeDefinition: function EnumTypeDefinition(_ref28) {
-	    var name = _ref28.name,
-	        directives = _ref28.directives,
-	        values = _ref28.values;
-	    return join(['enum', name, join(directives, ' '), block(values)], ' ');
-	  },
-
-	  EnumValueDefinition: function EnumValueDefinition(_ref29) {
-	    var name = _ref29.name,
-	        directives = _ref29.directives;
-	    return join([name, join(directives, ' ')], ' ');
-	  },
-
-	  InputObjectTypeDefinition: function InputObjectTypeDefinition(_ref30) {
-	    var name = _ref30.name,
-	        directives = _ref30.directives,
-	        fields = _ref30.fields;
-	    return join(['input', name, join(directives, ' '), block(fields)], ' ');
-	  },
-
-	  TypeExtensionDefinition: function TypeExtensionDefinition(_ref31) {
-	    var definition = _ref31.definition;
-	    return 'extend ' + definition;
-	  },
-
-	  DirectiveDefinition: function DirectiveDefinition(_ref32) {
-	    var name = _ref32.name,
-	        args = _ref32.arguments,
-	        locations = _ref32.locations;
-	    return 'directive @' + name + wrap('(', join(args, ', '), ')') + ' on ' + join(locations, ' | ');
-	  }
-	};
-
-	/**
-	 * Given maybeArray, print an empty string if it is null or empty, otherwise
-	 * print all items together separated by separator if provided
-	 */
-	function join(maybeArray, separator) {
-	  return maybeArray ? maybeArray.filter(function (x) {
-	    return x;
-	  }).join(separator || '') : '';
-	}
-
-	/**
-	 * Given array, print each item on its own line, wrapped in an
-	 * indented "{ }" block.
-	 */
-	function block(array) {
-	  return array && array.length !== 0 ? indent('{\n' + join(array, '\n')) + '\n}' : '{}';
-	}
-
-	/**
-	 * If maybeString is not null or empty, then wrap with start and end, otherwise
-	 * print an empty string.
-	 */
-	function wrap(start, maybeString, end) {
-	  return maybeString ? start + maybeString + (end || '') : '';
-	}
-
-	function indent(maybeString) {
-	  return maybeString && maybeString.replace(/\n/g, '\n  ');
-	}
-
-/***/ },
-/* 2 */
-/***/ function(module, exports) {
-
-	'use strict';
-
-	Object.defineProperty(exports, "__esModule", {
-	  value: true
-	});
-	exports.visit = visit;
-	exports.visitInParallel = visitInParallel;
-	exports.visitWithTypeInfo = visitWithTypeInfo;
-	/**
-	 *  Copyright (c) 2015, Facebook, Inc.
-	 *  All rights reserved.
-	 *
-	 *  This source code is licensed under the BSD-style license found in the
-	 *  LICENSE file in the root directory of this source tree. An additional grant
-	 *  of patent rights can be found in the PATENTS file in the same directory.
-	 */
-
-	var QueryDocumentKeys = exports.QueryDocumentKeys = {
-	  Name: [],
-
-	  Document: ['definitions'],
-	  OperationDefinition: ['name', 'variableDefinitions', 'directives', 'selectionSet'],
-	  VariableDefinition: ['variable', 'type', 'defaultValue'],
-	  Variable: ['name'],
-	  SelectionSet: ['selections'],
-	  Field: ['alias', 'name', 'arguments', 'directives', 'selectionSet'],
-	  Argument: ['name', 'value'],
-
-	  FragmentSpread: ['name', 'directives'],
-	  InlineFragment: ['typeCondition', 'directives', 'selectionSet'],
-	  FragmentDefinition: ['name', 'typeCondition', 'directives', 'selectionSet'],
-
-	  IntValue: [],
-	  FloatValue: [],
-	  StringValue: [],
-	  BooleanValue: [],
-	  NullValue: [],
-	  EnumValue: [],
-	  ListValue: ['values'],
-	  ObjectValue: ['fields'],
-	  ObjectField: ['name', 'value'],
-
-	  Directive: ['name', 'arguments'],
-
-	  NamedType: ['name'],
-	  ListType: ['type'],
-	  NonNullType: ['type'],
-
-	  SchemaDefinition: ['directives', 'operationTypes'],
-	  OperationTypeDefinition: ['type'],
-
-	  ScalarTypeDefinition: ['name', 'directives'],
-	  ObjectTypeDefinition: ['name', 'interfaces', 'directives', 'fields'],
-	  FieldDefinition: ['name', 'arguments', 'type', 'directives'],
-	  InputValueDefinition: ['name', 'type', 'defaultValue', 'directives'],
-	  InterfaceTypeDefinition: ['name', 'directives', 'fields'],
-	  UnionTypeDefinition: ['name', 'directives', 'types'],
-	  EnumTypeDefinition: ['name', 'directives', 'values'],
-	  EnumValueDefinition: ['name', 'directives'],
-	  InputObjectTypeDefinition: ['name', 'directives', 'fields'],
-
-	  TypeExtensionDefinition: ['definition'],
-
-	  DirectiveDefinition: ['name', 'arguments', 'locations']
-	};
-
-	var BREAK = exports.BREAK = {};
-
-	/**
-	 * visit() will walk through an AST using a depth first traversal, calling
-	 * the visitor's enter function at each node in the traversal, and calling the
-	 * leave function after visiting that node and all of its child nodes.
-	 *
-	 * By returning different values from the enter and leave functions, the
-	 * behavior of the visitor can be altered, including skipping over a sub-tree of
-	 * the AST (by returning false), editing the AST by returning a value or null
-	 * to remove the value, or to stop the whole traversal by returning BREAK.
-	 *
-	 * When using visit() to edit an AST, the original AST will not be modified, and
-	 * a new version of the AST with the changes applied will be returned from the
-	 * visit function.
-	 *
-	 *     const editedAST = visit(ast, {
-	 *       enter(node, key, parent, path, ancestors) {
-	 *         // @return
-	 *         //   undefined: no action
-	 *         //   false: skip visiting this node
-	 *         //   visitor.BREAK: stop visiting altogether
-	 *         //   null: delete this node
-	 *         //   any value: replace this node with the returned value
-	 *       },
-	 *       leave(node, key, parent, path, ancestors) {
-	 *         // @return
-	 *         //   undefined: no action
-	 *         //   false: no action
-	 *         //   visitor.BREAK: stop visiting altogether
-	 *         //   null: delete this node
-	 *         //   any value: replace this node with the returned value
-	 *       }
-	 *     });
-	 *
-	 * Alternatively to providing enter() and leave() functions, a visitor can
-	 * instead provide functions named the same as the kinds of AST nodes, or
-	 * enter/leave visitors at a named key, leading to four permutations of
-	 * visitor API:
-	 *
-	 * 1) Named visitors triggered when entering a node a specific kind.
-	 *
-	 *     visit(ast, {
-	 *       Kind(node) {
-	 *         // enter the "Kind" node
-	 *       }
-	 *     })
-	 *
-	 * 2) Named visitors that trigger upon entering and leaving a node of
-	 *    a specific kind.
-	 *
-	 *     visit(ast, {
-	 *       Kind: {
-	 *         enter(node) {
-	 *           // enter the "Kind" node
-	 *         }
-	 *         leave(node) {
-	 *           // leave the "Kind" node
-	 *         }
-	 *       }
-	 *     })
-	 *
-	 * 3) Generic visitors that trigger upon entering and leaving any node.
-	 *
-	 *     visit(ast, {
-	 *       enter(node) {
-	 *         // enter any node
-	 *       },
-	 *       leave(node) {
-	 *         // leave any node
-	 *       }
-	 *     })
-	 *
-	 * 4) Parallel visitors for entering and leaving nodes of a specific kind.
-	 *
-	 *     visit(ast, {
-	 *       enter: {
-	 *         Kind(node) {
-	 *           // enter the "Kind" node
-	 *         }
-	 *       },
-	 *       leave: {
-	 *         Kind(node) {
-	 *           // leave the "Kind" node
-	 *         }
-	 *       }
-	 *     })
-	 */
-	function visit(root, visitor, keyMap) {
-	  var visitorKeys = keyMap || QueryDocumentKeys;
-
-	  var stack = void 0;
-	  var inArray = Array.isArray(root);
-	  var keys = [root];
-	  var index = -1;
-	  var edits = [];
-	  var parent = void 0;
-	  var path = [];
-	  var ancestors = [];
-	  var newRoot = root;
-
-	  do {
-	    index++;
-	    var isLeaving = index === keys.length;
-	    var key = void 0;
-	    var node = void 0;
-	    var isEdited = isLeaving && edits.length !== 0;
-	    if (isLeaving) {
-	      key = ancestors.length === 0 ? undefined : path.pop();
-	      node = parent;
-	      parent = ancestors.pop();
-	      if (isEdited) {
-	        if (inArray) {
-	          node = node.slice();
-	        } else {
-	          var clone = {};
-	          for (var k in node) {
-	            if (node.hasOwnProperty(k)) {
-	              clone[k] = node[k];
-	            }
-	          }
-	          node = clone;
-	        }
-	        var editOffset = 0;
-	        for (var ii = 0; ii < edits.length; ii++) {
-	          var editKey = edits[ii][0];
-	          var editValue = edits[ii][1];
-	          if (inArray) {
-	            editKey -= editOffset;
-	          }
-	          if (inArray && editValue === null) {
-	            node.splice(editKey, 1);
-	            editOffset++;
-	          } else {
-	            node[editKey] = editValue;
-	          }
-	        }
-	      }
-	      index = stack.index;
-	      keys = stack.keys;
-	      edits = stack.edits;
-	      inArray = stack.inArray;
-	      stack = stack.prev;
-	    } else {
-	      key = parent ? inArray ? index : keys[index] : undefined;
-	      node = parent ? parent[key] : newRoot;
-	      if (node === null || node === undefined) {
-	        continue;
-	      }
-	      if (parent) {
-	        path.push(key);
-	      }
-	    }
-
-	    var result = void 0;
-	    if (!Array.isArray(node)) {
-	      if (!isNode(node)) {
-	        throw new Error('Invalid AST Node: ' + JSON.stringify(node));
-	      }
-	      var visitFn = getVisitFn(visitor, node.kind, isLeaving);
-	      if (visitFn) {
-	        result = visitFn.call(visitor, node, key, parent, path, ancestors);
-
-	        if (result === BREAK) {
-	          break;
-	        }
-
-	        if (result === false) {
-	          if (!isLeaving) {
-	            path.pop();
-	            continue;
-	          }
-	        } else if (result !== undefined) {
-	          edits.push([key, result]);
-	          if (!isLeaving) {
-	            if (isNode(result)) {
-	              node = result;
-	            } else {
-	              path.pop();
-	              continue;
-	            }
-	          }
-	        }
-	      }
-	    }
-
-	    if (result === undefined && isEdited) {
-	      edits.push([key, node]);
-	    }
-
-	    if (!isLeaving) {
-	      stack = { inArray: inArray, index: index, keys: keys, edits: edits, prev: stack };
-	      inArray = Array.isArray(node);
-	      keys = inArray ? node : visitorKeys[node.kind] || [];
-	      index = -1;
-	      edits = [];
-	      if (parent) {
-	        ancestors.push(parent);
-	      }
-	      parent = node;
-	    }
-	  } while (stack !== undefined);
-
-	  if (edits.length !== 0) {
-	    newRoot = edits[edits.length - 1][1];
-	  }
-
-	  return newRoot;
-	}
-
-	function isNode(maybeNode) {
-	  return maybeNode && typeof maybeNode.kind === 'string';
-	}
-
-	/**
-	 * Creates a new visitor instance which delegates to many visitors to run in
-	 * parallel. Each visitor will be visited for each node before moving on.
-	 *
-	 * If a prior visitor edits a node, no following visitors will see that node.
-	 */
-	function visitInParallel(visitors) {
-	  var skipping = new Array(visitors.length);
-
-	  return {
-	    enter: function enter(node) {
-	      for (var i = 0; i < visitors.length; i++) {
-	        if (!skipping[i]) {
-	          var fn = getVisitFn(visitors[i], node.kind, /* isLeaving */false);
-	          if (fn) {
-	            var result = fn.apply(visitors[i], arguments);
-	            if (result === false) {
-	              skipping[i] = node;
-	            } else if (result === BREAK) {
-	              skipping[i] = BREAK;
-	            } else if (result !== undefined) {
-	              return result;
-	            }
-	          }
-	        }
-	      }
-	    },
-	    leave: function leave(node) {
-	      for (var i = 0; i < visitors.length; i++) {
-	        if (!skipping[i]) {
-	          var fn = getVisitFn(visitors[i], node.kind, /* isLeaving */true);
-	          if (fn) {
-	            var result = fn.apply(visitors[i], arguments);
-	            if (result === BREAK) {
-	              skipping[i] = BREAK;
-	            } else if (result !== undefined && result !== false) {
-	              return result;
-	            }
-	          }
-	        } else if (skipping[i] === node) {
-	          skipping[i] = null;
-	        }
-	      }
-	    }
-	  };
-	}
-
-	/**
-	 * Creates a new visitor instance which maintains a provided TypeInfo instance
-	 * along with visiting visitor.
-	 */
-	function visitWithTypeInfo(typeInfo, visitor) {
-	  return {
-	    enter: function enter(node) {
-	      typeInfo.enter(node);
-	      var fn = getVisitFn(visitor, node.kind, /* isLeaving */false);
-	      if (fn) {
-	        var result = fn.apply(visitor, arguments);
-	        if (result !== undefined) {
-	          typeInfo.leave(node);
-	          if (isNode(result)) {
-	            typeInfo.enter(result);
-	          }
-	        }
-	        return result;
-	      }
-	    },
-	    leave: function leave(node) {
-	      var fn = getVisitFn(visitor, node.kind, /* isLeaving */true);
-	      var result = void 0;
-	      if (fn) {
-	        result = fn.apply(visitor, arguments);
-	      }
-	      typeInfo.leave(node);
-	      return result;
-	    }
-	  };
-	}
-
-	/**
-	 * Given a visitor instance, if it is leaving or not, and a node kind, return
-	 * the function the visitor runtime should call.
-	 */
-	function getVisitFn(visitor, kind, isLeaving) {
-	  var kindVisitor = visitor[kind];
-	  if (kindVisitor) {
-	    if (!isLeaving && typeof kindVisitor === 'function') {
-	      // { Kind() {} }
-	      return kindVisitor;
-	    }
-	    var kindSpecificVisitor = isLeaving ? kindVisitor.leave : kindVisitor.enter;
-	    if (typeof kindSpecificVisitor === 'function') {
-	      // { Kind: { enter() {}, leave() {} } }
-	      return kindSpecificVisitor;
-	    }
-	  } else {
-	    var specificVisitor = isLeaving ? visitor.leave : visitor.enter;
-	    if (specificVisitor) {
-	      if (typeof specificVisitor === 'function') {
-	        // { enter() {}, leave() {} }
-	        return specificVisitor;
-	      }
-	      var specificKindVisitor = specificVisitor[kind];
-	      if (typeof specificKindVisitor === 'function') {
-	        // { enter: { Kind() {} }, leave: { Kind() {} } }
-	        return specificKindVisitor;
-	      }
-	    }
-	  }
-	}
-
-/***/ }
-/******/ ]);
+var printer = require('./bundledPrinter');
+
+exports.print = function() {
+  console.warn(
+    "Warning - the `printer` exports from `graphql-tag` will be removed in the next major version." +
+    "\nSee https://github.com/apollographql/graphql-tag/issues/54 for more information."
+  );
+
+  return printer.print.apply(undefined, arguments);
+}
